requirements.txt
os
python-dotenv
tweepy
